requirements.txt
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.2
wandb
tqdm
python-dotenv
pandas